requirements.txt
Flask==2.0.1
redis==3.5.3
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
Pillow==8.3.2
opencv-python-headless==4.5.3.56